feat: Switch to using JSON Pointer strings
The driver for this is to be able to support yaml fields like
annotations:
  ingress.k8s.io/something: here

To update the value "here"  we need to be able to specify a key with
slashes and dots in it. It turns out there is an existing spec for
this called JSON Pointer. https://tools.ietf.org/html/rfc6901